+++ todo/src/navigation/Navigator.jsx
@@ -1,7 +1,7 @@
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { Image, ActivityIndicator, View, TouchableOpacity } from "react-native";
+import { Image, ActivityIndicator, View } from "react-native";
 //importing screens
 import Login from "../../src/screens/Login";
 import Home from "../../src/tabs/Home";
@@ -14,15 +14,31 @@ import home from "../../assets/home.png";
 import profile from "../../assets/profile.png";
 //import context
 import { AuthContext } from "../context/AuthContext";
-import { useContext } from "react";
-
-import AntDesign from "@expo/vector-icons/AntDesign";
+import { useContext,useEffect } from "react";
+import * as SecureStore from "expo-secure-store";
 
 const Navigator = () => {
   const Stack = createNativeStackNavigator();
 
   //token and loading state from context
-  const { token, loading } = useContext(AuthContext);
+  const { token, loading, setToken, setLoading } = useContext(AuthContext);
+
+  //To check login state for initial rendering
+  useEffect(() => {
+    const checktoken = async () => {
+      try {
+        const stored = await SecureStore.getItemAsync("token");
+        if (stored) {
+          setToken(stored);
+        }
+      } catch (e) {
+        console.log("Error fetching token", e);
+      } finally {
+        setLoading(false);
+      }
+    };
+    checktoken();
+  }, []);
 
   // Loader screen
   if (loading) {
@@ -39,6 +55,36 @@ const Navigator = () => {
       </View>
     );
   }
+
+  //for bottom tab navigation
+const TabNavigator = () => {
+  const Tab = createBottomTabNavigator();
+
+  return (
+    <Tab.Navigator initialRouteName="Home">
+      <Tab.Screen
+        name="Home"
+        component={Home}
+        options={{
+          tabBarIcon: () => (
+            <Image source={home} style={{ height: 30, width: 30 }} />
+          ),
+          tabBarActiveTintColor: "#DA70D6",
+        }}
+      />
+      <Tab.Screen
+        name="Profile"
+        component={Profile}
+        options={{
+          tabBarIcon: () => (
+            <Image source={profile} style={{ height: 30, width: 30 }} />
+          ),
+          tabBarActiveTintColor: "#DA70D6",
+        }}
+      />
+    </Tab.Navigator>
+  );
+};
 
   return (
     <NavigationContainer>
@@ -75,32 +121,4 @@ const Navigator = () => {
 
 export default Navigator;
 
-//for bottom tab navigation
-const TabNavigator = () => {
-  const Tab = createBottomTabNavigator();
 
-  return (
-    <Tab.Navigator initialRouteName="Home">
-      <Tab.Screen
-        name="Home"
-        component={Home}
-        options={{
-          tabBarIcon: () => (
-            <Image source={home} style={{ height: 30, width: 30 }} />
-          ),
-          tabBarActiveTintColor: "#DA70D6",
-        }}
-      />
-      <Tab.Screen
-        name="Profile"
-        component={Profile}
-        options={{
-          tabBarIcon: () => (
-            <Image source={profile} style={{ height: 30, width: 30 }} />
-          ),
-          tabBarActiveTintColor: "#DA70D6",
-        }}
-      />
-    </Tab.Navigator>
-  );
-};
